+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ADDED:
 - new class DocumentFactory that provides factory methods for creating Document objects
 - new factory methods for TransformerFactory for provide needed attributes
 - new factory methods for Transformer for provide needed attributes and output properties
+- new extension method for print a Document object as String object
 
 Version 1.1
 -------------